Fixed database error when deleting comment

+++ app/helper/RTMediaCommentNotification.php
@@ -101,9 +101,14 @@ class RTMediaCommentNotification extends RTMediaNotification {
      * @param   int     $comment_id
      */
     function remove_comment_notification($comment_id) {
-        $comment_notification_id = (int) get_comment_meta($comment_id, 'comment_notification_id', true);
-        BP_Notifications_Notification::delete(array('id' => $comment_notification_id));
-        delete_comment_meta($comment_id, 'comment_notification_id');
+        
+        $comment_notification_id = (int) get_comment_meta( $comment_id, 'comment_notification_id', true );
+        
+        if ( $comment_notification_id > 0 ) {
+            BP_Notifications_Notification::delete( array( 'id' => $comment_notification_id ) );
+            delete_comment_meta( $comment_id, 'comment_notification_id' );
+        }
+        
     }
 
 }
